% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{myfun}
\alias{myfun}
\title{A function}
\usage{
myfun(str_with_nbsp = "x = 4")
}
\arguments{
\item{str_with_nbsp}{A string, potentially with non-breaking spaces}
}
\value{
Nothing relevant
}
\description{
A function
}
